fix: adding test cases and todos for future enhancements
