+++ java/vector/src/main/java/org/apache/arrow/vector/validate/ValidateVectorBufferVisitor.java
@@ -51,7 +51,8 @@ public class ValidateVectorBufferVisitor implements VectorVisitor<Void, Void> {
 
     if (vector instanceof FieldVector) {
       FieldVector fieldVector = (FieldVector) vector;
-      int typeBufferCount = TypeLayout.getTypeBufferCount(arrowType, fieldVector);
+      // TODO: update the `getTypeBufferCount` method to use the `vector` (FieldVector).
+      int typeBufferCount = TypeLayout.getTypeBufferCount(arrowType);
       validateOrThrow(fieldVector.getFieldBuffers().size() == typeBufferCount,
           "Expected %s buffers in vector of type %s, got %s.",
               typeBufferCount, vector.getField().getType().toString(), fieldVector.getFieldBuffers().size());
